# 05/results/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: RESTful API Design
...
paths:
...
  /tour:
    post:
      tags:
        - tour
      summary: Add a new tour to the store
      description: ''
      operationId: postTour
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tour'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tour'
          links:
            Self:
              $ref: '#/components/links/PostTour'
            GetTours:
              $ref: '#/components/links/GetTours'
            GetTourById:
              $ref: '#/components/links/GetTourById'
        '405':
          description: Invalid input
          links:
            Self:
              $ref: '#/components/links/PostTour'
    get:
      tags:
        - tour
      summary: Get all tours
      description: ''
      operationId: getTours
      parameters: []
      responses:
        '200':
          description: Tours returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tour'
          links:
            Self:
              $ref: '#/components/links/GetTours'
            PostCountry:
              $ref: '#/components/links/PostTour'
  '/tour/{tourId}':
    put:
      tags:
        - tour
      summary: Put tour to the store
      description: ''
      operationId: putTour
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tour'
      parameters:
        - name: tourId
          in: path
          description: ID of tour that needs to be updated
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tour'
          links:
            Self:
              $ref: '#/components/links/GetTourById'
            GetTours:
              $ref: '#/components/links/GetTours'
        '405':
          description: Invalid input
          links:
            GetTours:
              $ref: '#/components/links/GetTours'
    get:
      tags:
        - tour
      summary: Get tour
      description: ''
      operationId: getTour
      parameters:
        - name: tourId
          in: path
          description: ID of tour to get
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tour returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tour'
          links:
            Self:
              $ref: '#/components/links/GetTourById'
            Update:
              $ref: '#/components/links/UpdateTourById'
            Delete:
              $ref: '#/components/links/DeleteTourById'
    delete:
      tags:
        - tour
      summary: Deletes a tour
      description: ''
      operationId: deleteTour
      parameters:
        - name: tourId
          in: path
          description: Tour id to delete
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tour deleted
          links:
            GetTours:
              $ref: '#/components/links/GetTours'
        '400':
          description: Invalid ID supplied
        '404':
          description: Tour not found
...
components:
  schemas:
    Country:
      properties:
        id:
          type: integer
        name:
          type: string
    Location:
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          $ref: '#/components/schemas/Country'
    Tour:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        customers:
          items:
            $ref: '#/components/schemas/Customer'
          type: array
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
  links:
...
    PostTour:
      operationId: postTour
    GetTours:
      operationId: getTours
    GetTourById:
      operationId: getTour
      parameters:
        tourId: '$response.body#/id'
    UpdateTourById:
      operationId: putTour
      parameters:
        tourId: '$response.body#/id'
    DeleteTourById:
      operationId: deleteTour
      parameters:
        tourId: '$response.body#/id'
...
servers:
  - url: 'http://t-a.org'
    variables: {}